Implement code signatures
- Update publish-localartifact:
- Streamline
- check if signature specified is usable
- Begin signing things (notably, the file catalog)

Still a work in progress. In particular, we must evaluate what we get (assembly nupkg, ps module, something else) and then act accordingly

We'll probably need to refactor too. Publish-"anything" isn't quite the best fit for a name, even if it was born out of necessity at the time.